Sólo se para constellation cuando es un nodo general.

+++ scripts/stop.sh
@@ -2,7 +2,12 @@
 
 set -e
 
-pkill -f constellation-node
+mapfile -t NODE_TYPE <~/alastria/data/NODE_TYPE
+
+if [[ "$NODE_TYPE" == "general" ]]; then
+    pkill -f constellation-node
+fi
+
 pkill -f geth
 
 set +e
